Add files via upload
Added comments. Added font name and scale in menu (now Options/Settings). Fixed bugs in site.css. Fixed (lots of) bugs in edition of codes and indices. webkit does not allow to set a cursor when element is empty. Avoided this bug in by introducing a hairspace at the end when the code or index is empty to be able to set a cursor. Workaround to avoid the hairspace and delete when converting to number.